Add explicit Webpack configuration
Although package.json referenced Webpack, there was not actually a
configuration file for Webpack, and it could not be run
successfully. It appears that the code was published as-is.

I've now added a configuration file to pass the code through Babel,
using the current version of Babel. The module is still exported as
a default CommonJS module, but is now located in dist/main.js
instead of index.js.