Allow using TypeScript syntax
This allows files to be written with a `.ts` extension, and for
actual TypeScript syntax to be used inside those files.

Note, though, that at this point the types are not actually
actively checked (unless by your editor). All this does is
stripping the type definitions and parsing the .ts files as
Javascript.