Add the dashboard from datasister
We can't literally inject the module, since it determines the
current user in a different way (based on who's logged in rather
than on which Pod we're running), so this is currently a copy-paste
of most code in there, with a few import paths changed around.

Still, it at least shows where to copy changes from datasister
into, and vice versa: it maintains a similar overall structure, so
modifications made here can relatively easily be carried over to
datasister.