Separating dashboard and outline
No reason to reuse elements from the outline container that is used for displaying panes for a given subject.

This commit separates the logic for the two, and handles switching between them.

Not done any styling on cancel button, as I want to know how Tim wants to have it first.

This fixes https://github.com/solid/solid-panes/issues/154