Fix homepage link to files
Currently a visitor not logged in sees a link to "user X's files" at user X's pod root.  The link is currently relative which breaks when the databrowser is not served as a server-front end (e.g. as a stand-alone webapp or in Data Kitchen).  This changes it to a full absolute URL.